Fix two big bugs
(1) The algorithm used in Partitioner was intended to
use set-equality comparisons on the set of block indices
that each world is connected to. Using a list caused false
negatives and hence, sporadic incorrect responses.

(2) Implemented a "contains" function on models and used
it in Relation.generateGraph to address an OutOfDomainError
that would occur when the left model is smaller than the
right model.